fix lints and add hover and tooltip effects to rulegraph
